% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_limit_values.R
\name{correct_limit_values}
\alias{correct_limit_values}
\title{Correct values based on limit symbols}
\usage{
correct_limit_values(
  data,
  values,
  new_values,
  limits = NULL,
  under = c("<", "<="),
  over = c(">", ">=")
)
}
\arguments{
\item{data}{A dataframe.}

\item{values}{The column containing values to correct.}

\item{new_values}{The column name for corrected values.}

\item{limits}{OPTIONAL. The column name containing symbols. If this argument is left empty, the function will find
lesser than symbols within the \code{values} column.}

\item{under}{A vector containing the symbol(s) for lesser than. Defaults to "<" and "<=".}

\item{over}{A vector containing the symbol(s) for greater than. Defaults to ">" and ">=". Functionality for 'greater
than' symbols has not been implemented yet.}
}
\value{
A dataframe with the corrected column.
}
\description{
The \code{correct_limit_values} function finds symbols such as "<" or "<=" in the \code{values} column, or a column of choice
to correct values. By default, the function looks for lesser than symbols and will divide the corresponding value
by 2.
}
\examples{
# Sample dataframe
df <- data.frame(
  values = c("10", "<20", "30", "<40", "50")
)

# Applying the function
corrected_df <- correct_limit_values(data = df, values = "values", new_values = "corrected_values")

# Display the result
print(corrected_df)
}
