% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhdhv_colors.R
\name{rhdhv_color_palette}
\alias{rhdhv_color_palette}
\title{Use RHDHV colour schemes}
\usage{
rhdhv_color_palette(palette = "none", type = "v", display = F)
}
\arguments{
\item{palette}{The name of a palette. Use \code{display = T} to view available palettes.}

\item{type}{'v', 'nv', 'df'. Returns the HEX codes in a vector, a named vector or a dataframe of 2 columns.}

\item{display}{Set to TRUE to display available palettes.}
}
\value{
Returns a (named) vector or dataframe.
}
\description{
Use RHDHV colour schemes
}
\examples{
rhdhv_color_palette("KRW", "nv")
rhdhv_color_palette(display = T)
}
