% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Round numbers}
\usage{
round2(x, digits)
}
\arguments{
\item{x}{A number.}

\item{digits}{Number of digits to round to.}
}
\value{
A rounded number
}
\description{
Round numbers
}
\examples{
# Example usage
number <- 3.14159
rounded_number <- round2(number, 2)

# Display the result
print(rounded_number)
}
