% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_rhdhv.R
\name{theme_rhdhv}
\alias{theme_rhdhv}
\title{theme_rhdhv}
\usage{
theme_rhdhv(
  x_labels = "h",
  color_scheme = 1,
  grid_lines = "none",
  grid_linetype = "solid",
  text_size = "m",
  legend_position = "bottom"
)
}
\arguments{
\item{x_labels}{One of 'h' (horizontal), 'v' (vertical), 'd' (diagonal)}

\item{color_scheme}{A number, currently only 1}

\item{grid_lines}{One of 'none', 'both', 'x' (vertical), 'y' (horizontal).}

\item{grid_linetype}{One of GGPlot-supported linetypes (e.g., solid, dotted, etc.).}

\item{text_size}{One of 's' (small), 'm' (medium), 'l' (large).}

\item{legend_position}{One of bottom, left, right, top.}
}
\value{
a ggplot theme.
}
\description{
theme_rhdhv
}
\examples{
x <- c(1, 2, 3)
y <- c(1, 2, 3)

ggplot(aes(x = x, y = y)) +
  geom_point() +
  theme_rhdhv()
}
